Move param deserializers to separate module
